Add handling for a failed set-default to Sun
Summary:
If set-default fails, aborting should probably do a rollback, since we don't
really know if the version fully deployed.  The link that Mr. Gorilla posts in
this case would indeed do a rollback, but Sun wasn't previously picking up on
that.  Now it (he?) will.

This is to match D20590 in culture-cow.

Test Plan: Fingers crossed.

Reviewers: benjaminpollack, chris, csilvers

Reviewed By: chris, csilvers

Subscribers: jlfwong, alan, csilvers

Differential Revision: https://phabricator.khanacademy.org/D20960