dockerignore node_modules
Summary:
in case the developer npm runs this locally, we don't want to copy their
local node_modules into the container image on build (it probably
wouldn't break anything for most npm modules, but if they contain native
compiled code it definitely would, because a Darwin/Windows compiled
module could be copied into the image).

(this is an issue because we copy the entire directory, and by default
when run locally npm install puts node_modules in the same directory)

Test Plan:
in theory one should generate node_modules locally, rebuild the
container, and verify the files were not copied, however I'm confident
enough in this change that things will work as designed to YOLO it.

Reviewers: benkraft, benjaminpollack

Reviewed By: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D21140