move adapter selection into environment variables
Summary:
with this change, running the container without an environment will
default to shell mode, to make it easier to test locally.

in production, kubernetes environment variables will activate the slack
adapter (which only works with a corresponding token, which was provided
in kube environment already)

Test Plan:
The dev side behavior works fine (as we saw in screenhero session), the
part I want to verify is production, which will we only see upon our
next production deploy.  We *could* do a no-op deploy for this, but
since we have incoming changes to this service today, could also wait
for that and see it all in one go.

Reviewers: benkraft, benjaminpollack

Reviewed By: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D21157