remove author from package.json
Summary:
it was autogenerated by npm init, but I am not the sole maintainer of
this package, and people should just look at the contributors page in
github or do a git blame if they want to know who to talk to.

in theory a `maintainers` field could be a valuable source of truth, but
only if it was a standard across the company with all our repos, and we
don't use NPM for most of them, so it's not a good solution for us!

Test Plan: None

Reviewers: benkraft, benjaminpollack

Reviewed By: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D21165