Pin to a different commit of hubot-slack
Letting slack try to reconnect if the client drops turns out to cause us a
number of issues, where Hubot keeps running (so kube doesn't notice or care),
but is connected to slack 0 or 2 times.  We could pin on the commit before
that, but this fork actually reverts with some slight improvements (telling
slack not to even try to auto-reconnect, and exiting with status 1), so we'll
use that instead.  If/when slackhq/hubot-slack #215 gets merged, we can go back
to pinning on their master.

Test plan: deploy it, and wait.

Auditors: mroth, benjaminpollack